chore: separated zustand stores
